add property UserId in Post

+++ Project.Hairdresser.Api/Hairdresser.Api/Controllers/V1/PostsController.cs
@@ -2,6 +2,7 @@
 using Hairdresser.Api.Contracts.Responses;
 using Hairdresser.Api.Contracts.V1;
 using Hairdresser.Api.Domain;
+using Hairdresser.Api.Extensions;
 using Hairdresser.Api.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
@@ -29,7 +30,11 @@ namespace Hairdresser.Api.Controllers.V1
         [HttpPost(ApiRoutes.Post.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest request) 
         {
-            var post = new Post { Name = request.Name };
+            var post = new Post {
+                Name = request.Name,
+                AccountId = Guid.Parse( HttpContext.GetUserId())
+            };
+
             if (post.Id != Guid.Empty)
             {
                 post.Id = Guid.NewGuid();
@@ -61,9 +66,16 @@ namespace Hairdresser.Api.Controllers.V1
         [HttpDelete(ApiRoutes.Post.Delete)]
         public async Task<IActionResult> Delete([FromRoute]Guid postId) 
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You don't own this post" });
+            }
+
             var delete = await _postService.DeletePostAsync(postId);
             if (delete)
                 return NoContent();
+
             return NotFound();
         }
 
@@ -71,11 +83,15 @@ namespace Hairdresser.Api.Controllers.V1
         [HttpPut(ApiRoutes.Post.Update)]
         public async Task<IActionResult> Update([FromRoute]Guid postId, [FromBody] UpdatePostRequest request)
         {
-            var post = new Post
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if(!userOwnsPost)
             {
-                Id = postId,
-                Name = request.Name,
-            };
+                return BadRequest(new { error = "You don't own this post" });
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
+
             var update = await _postService.UpdatePostAsync(post); 
             if(update)
                 return Ok();
